Gutted add method for parser
I removed the add method as it became clear it was unnecessary for this class.

The lowPriority method as well as a completely unimplemented "cleaning" method need to be worked on.

Current implementation of adding to the tree requires testing. A sample main of printing all the items via a post order method for ease of evaluation might be of use.